+++ src/commands/misc.h
@@ -13,5 +13,6 @@ void cmd_banlist (const char *);
 void cmd_exlist  (const char *);
 void cmd_ilist   (const char *);
 void cmd_who     (const char *);
+void cmd_rules   (const char *);
 
 #endif