+++ CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - Event 421 `ERR_UNKNOWNCOMMAND`
 - Event 444 `ERR_NOLOGIN`
 - Broadcast window activity for private messages
+- Command /rules
 
 ### Changed ###
 - On event join/part handle an empty user/host.